code re-structure
