+++ src/mutations/add_pool_info_db_pmtp.py
@@ -9,7 +9,7 @@ logger = setup_logger_util("add_pool_info_db_pmtp_mutation", formatter)
 
 
 def add_pool_info_db_pmtp_mutation(pool, height, native_asset_balance, external_asset_balance):
-    t0 = time()
+    t0 = time.time()
 
     sql_str = f"""
         delete from pmtp_pool_info where pool='{pool}'
@@ -51,5 +51,5 @@ def add_pool_info_db_pmtp_mutation(pool, height, native_asset_balance, external_
 
     database_service.execute_query(sql_str)
 
-    tf = time()
+    tf = time.time()
     logger.info(f"Pool Info updated in {tf-t0} seconds")
