+++ src/resolvers/add_pool_record_continuous.py
@@ -1,4 +1,4 @@
-from src.resolvers.add_liquidity_provider_pmtp import add_liquidity_provider_pmtp_resolver
+from src.resolvers.add_liquidity_provider import add_liquidity_provider_pmtp_resolver
 
 
 def add_pool_record_continuous_resolver():
